+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Unrelease
+
+Added
+  * AZ-1299: Add `application_group_id` output
+
 # v7.0.0 - 2023-10-06
 
 Added